change: make built_in comparators and functions as submodule
