+++ tests/api_server.py
@@ -5,7 +5,7 @@ from functools import wraps
 
 from flask import Flask, make_response, request
 
-from httprunner.built_in import gen_random_string
+from httprunner.builtin.functions import gen_random_string
 
 try:
     from httpbin import app as httpbin_app
